feat: refactor hive-utils crate to be part of test-utils crate
