--- crates/hive-utils/Cargo.toml
@@ -1,30 +1,0 @@
-[package]
-name = "hive-utils"
-version = { workspace = true }
-edition = { workspace = true }
-description = { workspace = true }
-homepage = { workspace = true }
-
-[dependencies]
-eyre = { workspace = true }
-kakarot-rpc-core = { path = "../core" }
-lazy_static = { workspace = true }
-reth-primitives = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-serde_with = { workspace = true }
-starknet = { workspace = true }
-starknet_api = { workspace = true }
-tokio = { workspace = true }
-url = { workspace = true }
-
-kakarot-test-utils = { path = "../test-utils" }
-
-[dev-dependencies]
-cargo-husky = { workspace = true }
-katana-core = { workspace = true }
-rstest = { workspace = true }
-
-[[bin]]
-name = "dump-genesis"
-path = "src/hive/bin/dump-genesis.rs"
