+++ crates/test-utils/src/hive_utils/kakarot.rs
@@ -27,12 +27,12 @@ pub fn compute_starknet_address(
 #[cfg(test)]
 mod tests {
     use kakarot_rpc_core::mock::constants::ACCOUNT_ADDRESS;
-    use kakarot_test_utils::deploy_helpers::KakarotTestEnvironmentContext;
-    use kakarot_test_utils::fixtures::kakarot_test_env_ctx;
     use rstest::*;
     use starknet::core::types::{BlockId, BlockTag};
 
     use super::compute_starknet_address;
+    use crate::deploy_helpers::KakarotTestEnvironmentContext;
+    use crate::fixtures::kakarot_test_env_ctx;
 
     /// This test is done against the Kakarot system deployed on the Starknet test sequencer.
     /// It tests the compute_starknet_address function by comparing the result of the computation
