+++ crates/test-utils/src/hive_utils/test_data/madara_genesis.json
