+++ crates/test-utils/src/lib.rs
@@ -2,3 +2,4 @@ pub mod constants;
 pub mod deploy_helpers;
 pub mod execution_helpers;
 pub mod fixtures;
+pub mod hive_utils;
